GH-127381: pathlib ABCs: remove `PathBase.stat()`
Remove the `PathBase.stat()` method. Its use of the `os.stat_result` API,
with its 10 mandatory fields and low-level types, makes it a poor fit for
virtual filesystems.

We'll look to add a `PathBase.info` attribute later - see GH-125413.
